requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.31.0
pymysql==1.1.0
boto3==1.34.14